requirements.txt
tensorflow-cpu>=2.5
Flask==2.0.2
gunicorn==20.1.0
Jinja2==3.0.2
numpy>=1.1
